# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AI4BMI API Groupe 3",
        "description": "API REST pour le système BMI — Gestion équipements et E-commerce BMI",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/api/v1/equipments": {
            "get": {
                "tags": [
                    "Équipements"
                ],
                "summary": "Lister tous les équipements",
                "description": "Retourne l'inventaire complet du parc matériel.",
                "operationId": "02b66b3bfecb572db7a59f3ab0857e03",
                "responses": {
                    "200": {
                        "description": "Inventaire récupéré avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Compresseur Atlas Copco"
                                            },
                                            "status": {
                                                "type": "string",
                                                "example": "actif"
                                            },
                                            "category": {
                                                "properties": {
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Air Comprimé"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Équipements"
                ],
                "summary": "Créer un équipement",
                "description": "Ajoute un nouvel équipement au système avec son statut initial.",
                "operationId": "97a2ab8c0d93114c8bef39babf18b858",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "equipment_category_id",
                                    "name",
                                    "serial_number"
                                ],
                                "properties": {
                                    "equipment_category_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Robot KUKA KR 210"
                                    },
                                    "brand": {
                                        "type": "string",
                                        "example": "KUKA"
                                    },
                                    "serial_number": {
                                        "type": "string",
                                        "example": "KUKA-003"
                                    },
                                    "installation_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-01-15"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "actif",
                                            "en_panne",
                                            "en_maintenance",
                                            "hors_service"
                                        ],
                                        "example": "actif"
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Zone A"
                                    },
                                    "picture": {
                                        "type": "string",
                                        "example": "equipments/robot.jpg"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Bras articulé pour soudure haute précision"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Équipement enregistré avec succès"
                    },
                    "403": {
                        "description": "Accès refusé — Privilèges insuffisants"
                    },
                    "422": {
                        "description": "Erreur de validation (ex: numéro de série déjà existant)"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/equipments/{id}": {
            "get": {
                "tags": [
                    "Équipements"
                ],
                "summary": "Voir les détails complets",
                "description": "Récupère un équipement avec tout son historique de maintenance et de pannes.",
                "operationId": "cf0b73415e1adbd817e7be46175c1514",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'équipement",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Fiche équipement récupérée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "maintenances": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "breakdowns": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Équipement introuvable"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Équipements"
                ],
                "summary": "Modifier un équipement",
                "description": "Met à jour les informations d'une machine existante.",
                "operationId": "2141a6f4eb5cf6d0702947ffec3826e7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "actif",
                                            "en_panne",
                                            "en_maintenance",
                                            "hors_service"
                                        ]
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Zone B"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Maintenance effectuée, RAS"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Équipement mis à jour avec succès"
                    },
                    "404": {
                        "description": "Équipement non trouvé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Équipements"
                ],
                "summary": "Supprimer un équipement",
                "description": "Retire définitivement l'équipement de la base de données.",
                "operationId": "1f395b34fedb037e112a2a1a0a32bbae",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Équipement supprimé avec succès"
                    },
                    "403": {
                        "description": "Action interdite — Admin requis"
                    },
                    "404": {
                        "description": "Équipement non trouvé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter token in format (Bearer <token>)"
            }
        }
    },
...
}